pyproject.toml
[tool.poetry]
name = "culture-questions-agent"
version = "0.1.0"
description = "Offline Hybrid RAG system for Cultural QA benchmark"
authors = ["Yahor Paromau <yahor.paromau@mailbox.tu-dresden.de>"]
readme = "README.md"
packages = [{include = "culture_questions_agent", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
hydra-core = "^1.3.2"
llama-index-core = "^0.14.12"
llama-index-llms-huggingface = "^0.6.1"
llama-index-embeddings-huggingface = "^0.6.1"
llama-index-readers-wikipedia = "^0.4.1"
llama-index-retrievers-bm25 = "^0.6.5"
transformers = "^4.57.3"
torch = "^2.9.1"
rank-bm25 = "^0.2.2"
pycountry = "^24.6.1"
tqdm = "^4.67.1"
wikipedia = "^1.4.0"
llama-index-llms-openai = "^0.6.12"
mlflow = "^3.8.1"
jinja2 = "^3.1.4"
pandas = "^2.3.3"
ipykernel = "^7.1.0"
keybert = "^0.9.0"
llama-index-tools-wikipedia = "^0.4.1"
llama-index-tools-google = "^0.6.2"
llama-index-tools-duckduckgo = "^0.4.1"
ddgs = "^9.10.0"
# NEW: Dependencies for SOTA RAG features
sentence-transformers = "^3.0.0"  # For semantic splitting and reranking
faiss-cpu = "^1.9.0"  # Optional: for efficient similarity search (use faiss-gpu on H100)
# faiss-gpu = "^1.9.0"  # Uncomment for GPU-accelerated FAISS
llama-index-postprocessor-flag-embedding-reranker = "^0.4.1"
llama-index-indices-managed-colbert = "^0.5.1"
colbert-ai = "^0.2.22"
flagembedding = "^1.3.5"
llama-index-vector-stores-qdrant = "^0.9.0"
trafilatura = "^2.0.0"
llama-index-readers-web = "^0.5.6"
llama-index-vector-stores-faiss = "^0.5.2"
llama-index-vector-stores-lancedb = "^0.4.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
